requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
pydantic>=2.0.0
python-dateutil>=2.8.0
pypdf>=4.0.0
python-docx>=1.0.0
openai>=1.30.0
python-dotenv>=1.0.0
